API

Public

CareerOneStop.CareerOneStop — Module

CareerOneStop::Module

API wraper for the Career One Stop API.

CareerOneStop.certificationfinder — Function

certificationfinder(keyword::AbstractString,
                    userId::AbstractString = ENV["API_CAREERONESTOP_USER"],
                    token::AbstractString = ENV["API_CAREERONESTOP"],
                    directFlag::AbstractString = "0",
                    industry::AbstractString = "0",
                    certType::AbstractString = "0",
                    organization::AbstractString = "0",
                    occupation::AbstractString = "0",
                    agency::AbstractString = "0",
                    sortColumn::AbstractString = "0",
                    sortDirections::AbstractString = "0",
                    startRecord::Integer = 0,
                    limitRecord::Integer = 10,
                    )

Consumes the List Certifications endpoint of the CareerOneStop API.

Example

julia> nurse = certificationfinder("nurse")
JSON3.Object{Vector{UInt8}, Vector{UInt64}} with 13 entries:
  :CertList                  => JSON3.Object[{…
  :CertAccredAgencyLegend    => JSON3.Object[{…
  :RecordCount               => 561
  :DidYouMean                => nothing
  :AutoCorrection            => nothing
  :IsDirectMatch             => "K"
  :OccupationCountList       => {…
  :IndustryCountList         => {…
  :TypeCountList             => {…
  :OrgCountList              => {…
  :RelatedCountList          => {…
  :CertDetail                => nothing
  :CertAccredAgencyCountList => {…

julia> @views nurse.CertList[1:2]
2-element view(::JSON3.Array{JSON3.Object, Vector{UInt8}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}, 1:2) with eltype JSON3.Object:
 {
                     "Id": "6126-C",
                   "Name": "Gerontological Nursing",
                    "Url": "https://www.nursingworld.org/our-certifications/gerontological-nurse/",
           "Organization": "ANA Enterprise",
                   "Type": "Specialty",
              "TypeOrder": "C",
                "Related": "",
   "CertAccredAgencyList": [
                             {
                                 "Name": "In-Demand",
                                "Order": "1"
                             },
                             {
                                 "Name": "NCCA",
                                "Order": "2"
                             },
                             {
                                 "Name": "MILITARY",
                                "Order": "4"
                             },
                             {
                                 "Name": "ABNS",
                                "Order": "7"
                             }
                           ],
    "OrganizationAddress": "8515 Georgia Avenue, Suite 400, Silver Spring, MD  20910-3492",
        "OrganizationUrl": "https://www.nursingworld.org/",
                "Acronym": "RN-BC",
            "Description": "The ANCC Gerontological Nursing board certification examination is a competency based examination that provides a valid and reliable assessment of the entry-level clinical knowledge and skills of registered nurses in the gerontological specialty after initial RN licensure. Once you complete eligibility requirements to take the certification examination and successfully pass the exam, you are awarded the credential: Registered Nurse-Board Certified (RN-BC).",
         "CertDetailList": [
                             {
                                 "Name": "More than two years of education or training after high school required?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "More than two years of work experience required?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Oral or Written Exam Required?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Renewal Required?",
                                "Value": "Every 5  Year(s)"
                             },
                             {
                                 "Name": "Renew through Continuing Educational Units(CEU)?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Renew through Re-Examination?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Renew through Continuing Professional Development(CPD)?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Does applicant have choice of at least two options from above for renewal (CEU, CPD, or exam)?",
                                "Value": "Yes"
                             }
                           ],
         "CertTestOption": [],
             "ExamDetail": nothing
}
 {
                     "Id": "6132-C",
                   "Name": "Pediatric Nursing",
                    "Url": "https://www.nursingworld.org/our-certifications/pediatric-nurse/",
           "Organization": "ANA Enterprise",
                   "Type": "Specialty",
              "TypeOrder": "C",
                "Related": "",
   "CertAccredAgencyList": [
                             {
                                 "Name": "In-Demand",
                                "Order": "1"
                             },
                             {
                                 "Name": "NCCA",
                                "Order": "2"
                             },
                             {
                                 "Name": "ANSI",
                                "Order": "3"
                             },
                             {
                                 "Name": "ABNS",
                                "Order": "7"
                             }
                           ],
    "OrganizationAddress": "8515 Georgia Avenue, Suite 400, Silver Spring, MD  20910-3492",
        "OrganizationUrl": "https://www.nursingworld.org/",
                "Acronym": "RN-BC",
            "Description": "The ANCC Pediatric Nursing board certification examination is a competency based examination that provides a valid and reliable assessment of the entry-level clinical knowledge and skills of registered nurses in the pediatric specialty after initial RN licensure. Once you complete eligibility requirements to take the certification examination and successfully pass the exam, you are awarded the credential: Registered Nurse-Board Certified (RN-BC).",
         "CertDetailList": [
                             {
                                 "Name": "More than two years of education or training after high school required?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "More than two years of work experience required?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Oral or Written Exam Required?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Renewal Required?",
                                "Value": "Every 5  Year(s)"
                             },
                             {
                                 "Name": "Renew through Continuing Educational Units(CEU)?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Renew through Re-Examination?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Renew through Continuing Professional Development(CPD)?",
                                "Value": "Yes"
                             },
                             {
                                 "Name": "Does applicant have choice of at least two options from above for renewal (CEU, CPD, or exam)?",
                                "Value": "Yes"
                             }
                           ],
         "CertTestOption": [],
             "ExamDetail": nothing
}

CareerOneStop.skillgap — Function

skillgap(onetCodeSource::AbstractString,
         onetCodeTarget::AbstractString,
         location::AbstractString = "US",
         userId::AbstractString = ENV["API_CAREERONESTOP_USER"],
         token::AbstractString = ENV["API_CAREERONESTOP"],
         radius::Integer = 25,
         )

Consumes the Skills Gaps endpoint of the CareerOneStop API.

Example

julia> rn_cns = skillgap("29114100", "29114104") # By default the location is 'US' which uses national data
JSON3.Object{Array{UInt8,1},Array{UInt64,1}} with 19 entries:
  :OccupationSkillsGapList   => JSON3.Object[{…
  :OccupationSkillsMatchList => JSON3.Object[{…
  :CurrentOccupationTitle    => "Registered Nurses"
  :TargetOccupationTitle     => "Clinical Nurse Specialists"
  :CurrentOccupationCode     => "29114100"
  :TargetOccupationCode      => "29114104"
  :CurrentOccupationWage     => "$73,300"
  :TargetOccupationWage      => "$73,300"
  :CurrentEducationTitle     => "Bachelor's degree"
  :TargetEducationTitle      => "Bachelor's degree"
  :CurrentExperienceTitle    => "None"
  :TargetExperienceTitle     => "None"
  :CurrentTrainingTitle      => "None"
  :TargetTrainingTitle       => "None"
  :CurrentLicenses           => 91
  :TargetLicenses            => 53
  :CurrentCertificates       => 169
  :TargetCertificates        => 45
  :LocationState             => "US"

julia> @views rn_cns.OccupationSkillsGapList[1:2]
2-element view(::JSON3.Array{JSON3.Object, Vector{UInt8}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}, 1:2) with eltype JSON3.Object:
 {
               "Title": "Education and Training",
         "Description": "Knowledge of principles and methods for curriculum and training design, teaching and instruction for individuals and groups, and the measurement of training effects.",
             "SkillId": "2.C.6",
      "SKScoreCurrent": "Medium",
       "SKScoreTarget": "Very High",
           "SkillType": "knowledge",
   "AvailableTraining": true
}
 {
               "Title": "Operations Analysis",
         "Description": "Figuring out what a product or service needs to be able to do.",
             "SkillId": "2.B.3.a",
      "SKScoreCurrent": "Very Low",
       "SKScoreTarget": "Medium",
           "SkillType": "skills",
   "AvailableTraining": false
}

julia> skillgap("29114101", "29114100", "VA")
JSON3.Object{Array{UInt8,1},Array{UInt64,1}} with 19 entries:
  :OccupationSkillsGapList   => nothing
  :OccupationSkillsMatchList => JSON3.Object[{…
  :CurrentOccupationTitle    => "Acute Care Nurses"
  :TargetOccupationTitle     => "Registered Nurses"
  :CurrentOccupationCode     => "29114101"
  :TargetOccupationCode      => "29114100"
  :CurrentOccupationWage     => "$70,200"
  :TargetOccupationWage      => "$70,200"
  :CurrentEducationTitle     => "Bachelor's degree"
  :TargetEducationTitle      => "Bachelor's degree"
  :CurrentExperienceTitle    => "None"
  :TargetExperienceTitle     => "None"
  :CurrentTrainingTitle      => "None"
  :TargetTrainingTitle       => "None"
  :CurrentLicenses           => 0
  :TargetLicenses            => 2
  :CurrentCertificates       => 40
  :TargetCertificates        => 169
  :LocationState             => "VA"

Private